curl golang sources with insecure cert option
I'm setting this as a workaround for a Docker problem. When a Docker
image starts up, the date is set to whenever the filesystem was
saved. Thus, I was getting an issue where curl was complaining that a
cert wasn't yet valid, since the site that hosts the golang source must
have recently rotated their certificate.